requirements.txt
alembic==1.6.5
beautifulsoup4==4.10.0
cachetools==4.2.0
pdfkit==0.6.1
Pillow==8.3.2
Pint==0.16
psycopg2-binary==2.8.6
pytz==2021.1
PyYAML==5.4
requests==2.26.0
Sanic-Cors==1.0.1
sanic-openapi==0.6.2
sanic-testing==0.7.0
sanic==21.9.1
SQLAlchemy==1.3.22
srt==3.4.1
websockets==10.1
webvtt-py==0.4.6
youtube-dl==2021.6.6
